requirements.txt
Django==5.2
django-filter
python-dotenv
psycopg